requirements.txt
flask
flask-socketio
requests
beautifulsoup4
pywebview
qrcode
selenium
webdriver-manager
Pillow